Update 1-Installation_and_configuration.md
You can install the Docker engine on Windows and Linux platforms. Since the container is using the Kernel of the host operating system, if we want to run a Windows container, we will need to install the Docker engine on a Windows server. If we install the Docker engine on a Linux host machine, any containers running there will use the Linux Kernel of the host operating system. Therefore, we will be able to exclusively run Linux containers on a Linux host machine.