Update 1-Installation_and_configuration.md
The approach that many customers use is to have all three manager nodes run Docker on a Linux machine and have two different groups of worker nodes: a Linux group and a Windows group. This way, they will be able to deploy Windows and Linux applications on their container platform.