Update 1-Installation_and_configuration.md
It is impossible to run a Windows container on a Linux worker. But it is possible to run a Linux container on a Windows server. The reason is that modern Windows servers have a Linux kernel embedded in the Windows operating system, which makes it possible to run both types of containers on a Windows machine: Linux and Windows containers. There are some features that are only available when running Linux containers on a Linux host machine. An example of such a feature is the ability to mount a volume in memory on a temporary file system.