Update 1-Installation_and_configuration.md
File system management and how Docker works under the hood is also different for Windows and Linux. It is true that Windows provides with the Compute Service an equivalent set of features to mimic Linux features related to file systems, control groups, namespaces, and other capabilities of the Linux operating system, but this Compute Service will never be exactly the same as its Linux counterpart because there are some Linux features that are simply impossible to reproduce on a Windows operating system.