Update 2-Namespaces_and_cgroups.md
Namespaces are partitions of Linux kernel resources such that a set of resources, such as the process identifier (PID), is assigned to a specific Linux namespace and can only be accessed from within that namespace. In this case, the PID namespaces are nested, which means that from the root namespace you will be able to see the PID of any other process running in a nested namespace but you won't see the PID of any other process running outside of its own namespace.