Update 2-Namespaces_and_cgroups.md
Control groups (cgroups) are another feature of the Linux kernel initially released in 2007. It allows for the isolation of resource usage by a set of processes. You can control CPU and memory usage, disk I/O and network performance, and many other resources. It is not mandatory to configure your containers with control groups, but Docker officially considers it a best practice. Otherwise, you may compromise the stability of your cluster in case your containers exhaust the hardware resources of the host machine.