Update 2-Namespaces_and_cgroups.md
It is also important to understand the fundamental difference between containers and virtual machines. A virtual machine is built on a set of virtual resources provided by virtualization software, which is the hypervisor. The hypervisor will provide access to the virtual resources created on top of the actual hardware of the host machine. In the virtual machine we will install a complete operating system including the kernel as well as our main application with all the necessary dependencies. A typical example would be a virtual machine with a full operating system plus a Tomcat, MySQL, or Apache web server.