Update 2-Namespaces_and_cgroups.md
Namespaces ensure the isolation of some operating system resources, such as file system mount points, hostname, network stack, or PID. But they do not limit the use of memory, disk or CPU. To control the use of physical resources, we need to configure the control groups for that container. It is strongly recommended to configure control groups for any container that will be used in a production environment.