Update 2-Namespaces_and_cgroups.md
This feature was available in the Linux kernel many years before Docker was developed. This set of namespaces and control groups is what we call a "container". The container assigned by default to any process running on a Linux machine is also called the global or root container. Docker uses this same technology to configure local Docker containers (typically nested namespaces of the global root namespace).